+++ src/App.tsx
@@ -1,23 +1,23 @@
-import { BaseButton } from './common/components/basebutton';
+import { Button } from './common/components/button';
 import { KakaoAuthButton } from './common/components/kakaoauthbutton';
 
 function App() {
   return (
     <div>
       {/* 기본 디폴트 테마 사용 */}
-      <BaseButton>Default</BaseButton>
+      <Button>Default</Button>
 
       {/* 디자인 시스템 테마 주입 */}
-      <BaseButton variant="destructive" size="sm">
+      <Button variant="destructive" size="sm">
         Small + Destructive
-      </BaseButton>
+      </Button>
 
       {/* 디자인 시스템 테마 & asChild에 따른 Slot 사용 */}
-      <BaseButton variant="link" asChild>
+      <Button variant="link" asChild>
         <a href="https://example.com" target="_blank" rel="noreferrer">
           External Link
         </a>
-      </BaseButton>
+      </Button>
 
       {/* 내부적으로 고정된 버튼 스타일 및 이벤트 사용 */}
       <KakaoAuthButton>카카오 로그인</KakaoAuthButton>
