+++ src/common/components/kakaoauthbutton/index.tsx
@@ -1,13 +1,13 @@
 import { forwardRef } from 'react';
 
-import { Button, ButtonProps } from '../button';
+import { BaseButton, BaseButtonProps } from '../basebutton';
 
 import * as styles from './KakaoAuthButton.styles';
 
-/** 고정된 로직과 스타일 사용*/
-const KakaoAuthButton = forwardRef<HTMLButtonElement, ButtonProps>((props, ref) => {
+/** 고정된 스타일과 로직 사용 */
+const KakaoAuthButton = forwardRef<HTMLButtonElement, BaseButtonProps>((props, ref) => {
   return (
-    <Button
+    <BaseButton
       onClick={() => console.log('Kakao login!')}
       css={styles.kakaoButtonStyle}
       ref={ref}
